+++ Assets/Source/Modules/ScoreFlyActions/IScoring.cs
@@ -3,5 +3,5 @@ public interface IScoring
     public void Init(IScorePresenter presenter, IScoreConvention scoreConvention);
     public void StartScoring();
     public void StopScoring();
-    public float GetCurrentScore();
+    public int GetCurrentScore();
 }
